Added a cardlibrary and the forestfire event
In order to implement the forest fire event, a card library was necessary. Meaning a choice of cards out of a subset of cards of the current deck of the player with which several things can be done. In this event, card removal from the deck. (Implemented a function for exchanging a card for a random new one aswell)