Sphere stuck handling and major bug fixes for orbs
So much was wrong with orbs due to refactor, but I fixed most problems with it. Spheres now disappear and start next phase if stuck for 2 seconds. Be watchful if it works okay for the clone sphere, I did think about it and it should work, but you never know.